refactor: separate internal KV and HTTP utilities
