+++ lib/handle_callback_test.ts
@@ -1,11 +1,8 @@
 // Copyright 2023 the Deno authors. All rights reserved. MIT license.
 import { handleCallback } from "./handle_callback.ts";
 import { assertEquals, assertRejects, returnsNext, stub } from "../dev_deps.ts";
-import {
-  getAndDeleteOAuthSession,
-  OAUTH_COOKIE_NAME,
-  setOAuthSession,
-} from "./_core.ts";
+import { getAndDeleteOAuthSession, setOAuthSession } from "./_kv.ts";
+import { OAUTH_COOKIE_NAME } from "./_http.ts";
 import {
   assertRedirect,
   randomOAuthConfig,
