+++ lib/sign_in_test.ts
@@ -1,7 +1,8 @@
 // Copyright 2023 the Deno authors. All rights reserved. MIT license.
 import { signIn } from "./sign_in.ts";
 import { assertEquals, assertNotEquals, getSetCookies } from "../dev_deps.ts";
-import { getAndDeleteOAuthSession, OAUTH_COOKIE_NAME } from "./_core.ts";
+import { OAUTH_COOKIE_NAME } from "./_http.ts";
+import { getAndDeleteOAuthSession } from "./_kv.ts";
 import { assertRedirect, randomOAuthConfig } from "./_test_utils.ts";
 
 Deno.test("signIn() returns a response that signs-in the user", async () => {
