+++ lib/sign_out_test.ts
@@ -1,7 +1,7 @@
 // Copyright 2023 the Deno authors. All rights reserved. MIT license.
 import { assertEquals } from "../dev_deps.ts";
 import { signOut } from "./sign_out.ts";
-import { SITE_COOKIE_NAME } from "./_core.ts";
+import { SITE_COOKIE_NAME } from "./_http.ts";
 import { assertRedirect } from "./_test_utils.ts";
 
 Deno.test("signOut() returns a redirect response if the user is not signed-in", async () => {
